向团队中添加成员

+++ WT-service/WT-company-service/src/main/java/cn/edu/szu/company/pojo/domain/UserGroupRequest.java
@@ -11,5 +11,5 @@ import java.util.List;
 @AllArgsConstructor
 public class UserGroupRequest {
     private List<String> list;
-    private int groupId;
+    private Long groupId;
 }
